README
Gomoku (Connect 4) is a game where the user gets to pick a location in a 2-D array
to put their game piece, the next user then takes their turn, and the cycle continues
until one of the players has connected four of their pieces in a row. In this implementation
the game holds memory of a scoreboard consisting of wins losses and ties for each of the 
players.